Create marketing branch
Created marketing branch, some classes, RDMBS connection in the pom.xml and preliminary interfaces for the marketing team, alongside comments for TODO.
