+++ lancasterhall/src/main/java/in2033/Main.java
@@ -1,0 +1,7 @@
+package in2033;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
