+++ src/databaseconnector.java
@@ -1,0 +1,6 @@
+public class databaseconnector {
+    public void connectAndQuery(String user, String pass)
+    {
+        
+    }
+}
